--- a/doc/source/whatsnew/v1.3.3.rst
+++ b/doc/source/whatsnew/v1.3.3.rst
@@ -25,7 +25,7 @@ Fixed regressions
 
 Performance improvements
 ~~~~~~~~~~~~~~~~~~~~~~~~
-- Performance improvement for :meth:`DataFrame.__setitem__` with dataframes containing unique columns and list-like values
+- Performance improvement for :meth:`DataFrame.__setitem__` with dataframes containing unique columns
 -
 -
 .. ---------------------------------------------------------------------------