--- a/doc/source/whatsnew/v1.3.3.rst
+++ b/doc/source/whatsnew/v1.3.3.rst
@@ -28,6 +28,7 @@ Performance improvements
 - Performance improvement for :meth:`DataFrame.__setitem__` when the key or value is not a :class:`DataFrame`, or key is not list-like (:issue:`43274`)
 -
 -
+
 .. ---------------------------------------------------------------------------
 
 .. _whatsnew_133.bug_fixes: