@jreback 2021-08-28
can u run the indexing asvs and report here