@mzeitlin11 2021-08-28
Could also go in 1.3.3, looks like this helps top regression here: https://simonjayhawkins.github.io/fantastic-dollop/#regressions?sort=3&dir=desc&branch=1.3.x