@phofl 2021-08-28
Yeah good point, this check was added in 1.3