@debnathshoham 2021-08-28
should I put the whatsnew in 1.3.3 fixed regression or under others?
@jreback - I have added the indexing asvs report in the top comment