@simonjayhawkins 2021-08-29
> ps://simonjayhawkins.github.io/fantastic-dollop/#regressions?sort=3&dir=desc&branch=1.3.x

side note: `class InsertColumns` was updated in #42998. Changing the code or setup code invalidates previous results for that benchmark. So that regression is not currently reported.