@simonjayhawkins 2021-08-30
> side note: `class InsertColumns` was updated in #42998. Changing the code or setup code invalidates previous results for that benchmark. So that regression is not currently reported.

benchmarks have been re-run and gh-pages updated.